+++ js/util.js
@@ -1,0 +1,10 @@
+function getRandomInteger(min, max) {
+  return Math.floor(Math.random()
+    * (max - min + 1)) + min;
+}
+
+function getRandomArrayElement(elements) {
+  return elements[getRandomInteger(0, elements.length - 1)];
+}
+
+export { getRandomInteger, getRandomArrayElement };
